Code that makes -s on startup.sh work properly.

+++ startup.sh
@@ -22,21 +22,25 @@ if [ "$1" = "-n" -o "$1" = "--no-sleep" -o "$1" = "--no-delay" ]
 then
 	MSG=""
 	DELAY=""
+	shift
 fi
 if [ "$1" = "-d" -o "$1" = "--delay" ]
 then
 	MSG="echo Waiting $2 seconds before starting"
 	DELAY="sleep $2"
+	shift
+	shift
 fi
 if [ "$1" = "-m" -o "$1" = "--message-delay" ]
 then
 	MSG="echo Waiting $2 seconds for response before starting"
 	#DELAY="xmessage -buttons Now:0,Cancel:1 -default Now -timeout $2 Starting PiClock in $2 seconds"
 	DELAY='zenity --question --title PiClock --ok-label=Now --cancel-label=Cancel --timeout '$2' --text "Starting PiClock in '$2' seconds" >/dev/null 2>&1'
+	shift
+	shift
 fi
 
 $MSG
-echo $DELAY
 eval $DELAY
 if [ $? -eq 1 ]
 then
@@ -109,15 +113,20 @@ fi
 
 # the main app
 cd Clock
-# create a new log file name, max of 7 log files
-echo "Rotating log files...."
-rm PyQtPiClock.7.log >/dev/null 2>&1
-mv PyQtPiClock.6.log PyQtPiClock.7.log >/dev/null 2>&1
-mv PyQtPiClock.5.log PyQtPiClock.6.log >/dev/null 2>&1
-mv PyQtPiClock.4.log PyQtPiClock.5.log >/dev/null 2>&1
-mv PyQtPiClock.3.log PyQtPiClock.4.log >/dev/null 2>&1
-mv PyQtPiClock.2.log PyQtPiClock.3.log >/dev/null 2>&1
-mv PyQtPiClock.1.log PyQtPiClock.2.log >/dev/null 2>&1
-# start the clock
-echo "Starting PiClock...."
-python -u PyQtPiClock.py >PyQtPiClock.1.log 2>&1
+if [ "$1" = "-s" -o "$1" = "--screen-log" ]
+then
+  echo "Starting PiClock.... logging to screen."
+  python -u PyQtPiClock.py
+else
+  # create a new log file name, max of 7 log files
+  echo "Rotating log files...."
+  rm PyQtPiClock.7.log >/dev/null 2>&1
+  mv PyQtPiClock.6.log PyQtPiClock.7.log >/dev/null 2>&1
+  mv PyQtPiClock.5.log PyQtPiClock.6.log >/dev/null 2>&1
+  mv PyQtPiClock.4.log PyQtPiClock.5.log >/dev/null 2>&1
+  mv PyQtPiClock.3.log PyQtPiClock.4.log >/dev/null 2>&1
+  mv PyQtPiClock.2.log PyQtPiClock.3.log >/dev/null 2>&1
+  mv PyQtPiClock.1.log PyQtPiClock.2.log >/dev/null 2>&1
+  echo "Starting PiClock.... logging to Clock/PyQtPiClock.1.log "
+  python -u PyQtPiClock.py >PyQtPiClock.1.log 2>&1
+fi
